+++ cmd/pg/copy.go
@@ -2,6 +2,7 @@ package pg
 
 import (
 	"github.com/spf13/cobra"
+	"github.com/wal-g/tracelog"
 	"github.com/wal-g/wal-g/internal"
 )
 
@@ -43,6 +44,10 @@ var (
 )
 
 func runBackupCopy(cmd *cobra.Command, args []string) {
+	if withoutHistory {
+		tracelog.InfoLogger.Fatal("Sorry but without-history flag not supported yet.")
+		return
+	}
 	internal.HandleCopy(fromConfigFile, toConfigFile, backupName)
 }
